refactored, original tests passing
